feat: navigate through pages faster
We don't need to replace an entire screen to navigate through pages in the admin panel.

Each page in admin folder wraps the admin wrapper (View/wrappers/admin), so it makes sense to send just the main content alone instead of an entire web page. This results faster Page Load.

AdminController will check a request header to determine if the request was made by clicking a navigation link or not.